Restructuring File Structure
