+++ GarbageCan/judgingCriteria.html
@@ -11,7 +11,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
             integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
             crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="cssFiles/judgingCriteria.css">
+    <link rel="stylesheet" href="../cssFiles/judgingCriteria.css">
 </head>
 
 <body>
