+++ GarbageCan/pastEventsPage.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Past Events Page</title>
-    <link rel="icon" type="image/x-icon" href="images/favicon.png">
+    <link rel="icon" type="image/x-icon" href="../images/favicon.png">
     <!-- CSS only -->
     <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css"
           integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" rel="stylesheet">
@@ -12,7 +12,7 @@
             integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
             crossorigin="anonymous"></script>
     <!Link to CSS file>
-    <link rel="stylesheet" href="cssFiles/pastEventsPage.css">
+    <link rel="stylesheet" href="../cssFiles/pastEventsPage.css">
 </head>
 <body>
 <!Dropdown menu - placed at the top left of screen>
