+++ GarbageCan/sponsorPage.html
@@ -2,8 +2,8 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Hardware Page</title>
-    <link rel="icon" type="image/x-icon" href="images/favicon.png">
+    <title>About our Sponsors</title>
+    <link rel="icon" type="image/x-icon" href="../images/favicon.png">
     <!-- CSS only -->
     <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css"
           integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" rel="stylesheet">
@@ -11,7 +11,8 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
             integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
             crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="cssFiles/hardwarePage.css">
+    <link rel="stylesheet" href="../cssFiles/sponsorPage.css">
+
 </head>
 <body>
 <!--NAVIGATION BAR-->
@@ -51,27 +52,29 @@
         </div>
     </div>
 </nav>
-
 <!Title of the page - aligned at the center>
 <div style="text-align: center; height: 50px;">
-    Hardware Page
+    Sponsor Page
 </div>
 <!Title of table - everything within this div can be edited through the page>
-<div class="hardwareInfo" contenteditable="true">
-    Hardware Information
+<div class="sponsorInfo" contenteditable="true">
+    Sponsor Information
     <!Table for now, most likely will change later>
     <table style="width: 50%; font-size: 25px;">
         <tr>
-            <th>Hardware Name</th>
-            <th>Availability</th>
+            <th>Company Name</th>
+            <th>Sponsor Name</th>
+            <th>Tier</th>
         </tr>
         <tr>
-            <td>Example Hardware Name 1</td>
-            <td>Example Availability 1</td>
+            <td>Example Company Name 1</td>
+            <td>Example Sponsor Name 1</td>
+            <td>Example Tier 1</td>
         </tr>
         <tr>
-            <td>Example Hardware Name 2</td>
-            <td>Example Availability 2</td>
+            <td>Example Company Name 2</td>
+            <td>Example Sponsor Name 2</td>
+            <td>Example Tier 2</td>
         </tr>
     </table>
 </div>
