+++ GarbageCan/welcomePage.html
@@ -10,8 +10,8 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
             integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
             crossorigin="anonymous"></script>
-    <link rel="icon" type="image/x-icon" href="images/favicon.png">
-    <link href="cssFiles/welcomePage.css" rel="stylesheet">
+    <link rel="icon" type="image/x-icon" href="../images/favicon.png">
+    <link href="../cssFiles/welcomePage.css" rel="stylesheet">
 
 </head>
