--- simple/config.mjs
@@ -1,5 +1,0 @@
-export default {
-  user: process.env.MYSQLX_USER,
-  password: process.env.MYSQLX_PASSWORD,
-  host: process.env.MYSQLX_HOST
-}
