+++ views/css/hardwarePage.css
