+++ views/hardware.pug
