+++ views/hardwarePage.pug
